Fix sqlx Pool API type mismatch in PoolStats
- Cast num_idle() and size() return values from usize to u32
- sqlx::Pool methods return usize, but PoolStats expects u32
- All 126 lib tests passing

+++ src/handlers/database/engines/postgresql/connection.rs
@@ -141,8 +141,8 @@ impl PostgreSqlPool {
     /// Get pool statistics
     pub fn stats(&self) -> PoolStats {
         PoolStats {
-            num_idle: self.pool.num_idle(),
-            size: self.pool.size(),
+            num_idle: self.pool.num_idle() as u32,
+            size: self.pool.size() as u32,
         }
     }
